pyproject.toml
[tool.poetry]
name = "koch11"
version = "0.1.0"
description = ""
authors = ["Sho Okumoto <lebesgue0118@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dynamixel-sdk = "^3.7.31"
tqdm = "^4.67.1"
numpy = "1.26.4"
scipy = "^1.15.1"
genesis-world = "^0.2.1"
torch = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
